+++ CHANGELOG.md
@@ -1,4 +1,13 @@
 # rakuda #
 
+
+## 1.1.0.pre ##
+
+Features:
+  - add `associations` option to Verify
+  - add `scope.joins` option to Verify
+  - add `scope.wheres` option to Verify
+  - add `scope.orders` option to Verify
+
 ## 1.0.0 ##
 * Release